add geoloader package.
