+++ geoloader/init.go
@@ -1,0 +1,125 @@
+package geoloader
+
+import (
+	"context"
+	"errors"
+	"log"
+	"regexp"
+	"time"
+
+	"cloud.google.com/go/storage"
+	"google.golang.org/api/iterator"
+
+	"github.com/m-lab/annotation-service/api"
+)
+
+const (
+	// This is the date we have the first GeoLite2 dataset.
+	// Any request earlier than this date using legacy binary datasets
+	// later than this date using GeoLite2 datasets
+	GeoLite2CutOffDate = "August 15, 2017"
+)
+
+// DatasetNames are list of datasets sorted in lexographical order in downloader bucket.
+var DatasetNames []string
+
+// The date of lastest available dataset.
+var LatestDatasetDate time.Time
+
+// This is the regex used to filter for which files we want to consider acceptable for using with Geolite2
+var GeoLite2Regex = regexp.MustCompile(`Maxmind/\d{4}/\d{2}/\d{2}/\d{8}T\d{6}Z-GeoLite2-City-CSV\.zip`)
+
+// This is the regex used to filter for which files we want to consider acceptable for using with legacy dataset
+var GeoLegacyRegex = regexp.MustCompile(`.*-GeoLiteCity.dat.*`)
+var GeoLegacyv6Regex = regexp.MustCompile(`.*-GeoLiteCityv6.dat.*`)
+
+// UpdateArchivedFilenames extracts the filenames from downloader bucket
+// It also searches the latest Geolite2 files available in GCS.
+// It will also set LatestDatasetDate as the date of lastest dataset.
+// If it encounters an error, it will halt the program.
+func UpdateArchivedFilenames() error {
+	ctx := context.Background()
+	client, err := storage.NewClient(ctx)
+	if err != nil {
+		return err
+	}
+	prospectiveFiles := client.Bucket(api.MaxmindBucketName).Objects(ctx, &storage.Query{Prefix: api.MaxmindPrefix})
+	filename := ""
+	for file, err := prospectiveFiles.Next(); err != iterator.Done; file, err = prospectiveFiles.Next() {
+		if err != nil {
+			return err
+		}
+		DatasetNames = append(DatasetNames, file.Name)
+		if file.Name > filename && GeoLite2Regex.MatchString(file.Name) {
+			filename = file.Name
+		}
+	}
+	if err != nil {
+		log.Println(err)
+	}
+
+	// Now set the lastest dataset
+	date, err := ExtractDateFromFilename(filename)
+	if err != nil {
+		log.Println(err)
+		return err
+	}
+	LatestDatasetDate = date
+	return nil
+}
+
+// ExtractDateFromFilename return the date for a filename like
+// gs://downloader-mlab-oti/Maxmind/2017/05/08/20170508T080000Z-GeoLiteCity.dat.gz
+// TODO move this to maxmind package
+// TODO - actually, this now seems to be dead code.  But probably needed again soon, so leaving it here.
+func ExtractDateFromFilename(filename string) (time.Time, error) {
+	re := regexp.MustCompile(`[0-9]{8}T`)
+	filedate := re.FindAllString(filename, -1)
+	if len(filedate) != 1 {
+		return time.Time{}, errors.New("cannot extract date from input filename")
+	}
+	return time.Parse(time.RFC3339, filedate[0][0:4]+"-"+filedate[0][4:6]+"-"+filedate[0][6:8]+"T00:00:00Z")
+}
+
+// SelectArchivedDataset returns the archived GelLite dataset filename given a date.
+// For any input date earlier than 2013/08/28, we will return 2013/08/28 dataset.
+// For any input date later than latest available dataset, we will return the latest dataset
+// Otherwise, we return the last dataset before the input date.
+func SelectArchivedDataset(requestDate time.Time) (string, error) {
+	earliestArchiveDate, _ := time.Parse("January 2, 2006", "August 28, 2013")
+	if requestDate.Before(earliestArchiveDate) {
+		return "Maxmind/2013/08/28/20130828T184800Z-GeoLiteCity.dat.gz", nil
+	}
+	CutOffDate, _ := time.Parse("January 2, 2006", GeoLite2CutOffDate)
+	lastFilename := ""
+	for _, fileName := range DatasetNames {
+		if requestDate.Before(CutOffDate) && (GeoLegacyRegex.MatchString(fileName) || GeoLegacyv6Regex.MatchString(fileName)) {
+			// search legacy dataset
+			fileDate, err := ExtractDateFromFilename(fileName)
+			if err != nil {
+				continue
+			}
+			// return the last dataset that is earlier than requestDate
+			if fileDate.After(requestDate) {
+				return lastFilename, nil
+			}
+			lastFilename = fileName
+		} else if !requestDate.Before(CutOffDate) && GeoLite2Regex.MatchString(fileName) {
+			// Search GeoLite2 dataset
+			fileDate, err := ExtractDateFromFilename(fileName)
+			if err != nil {
+				continue
+			}
+			// return the last dataset that is earlier than requestDate
+			if fileDate.After(requestDate) {
+				return lastFilename, nil
+			}
+			lastFilename = fileName
+		}
+	}
+	// If there is no filename selected, return the latest dataset
+	if lastFilename == "" {
+		return "", errors.New("cannot find proper dataset")
+	}
+	return lastFilename, nil
+}
