+++ handler/pubsub.go
@@ -36,7 +36,7 @@ func waitForDownloaderMessages() {
 	}
 	// Block forever to listen for new messages and run the refresh dataset callbacks when a new message arrives
 	log.Fatal(sub.Receive(context.Background(), func(ctx context.Context, m *pubsub.Message) {
-		manager.PopulateLatestData()
+		manager.InitDataset()
 		m.Ack()
 	}))
 }
