--- manager/init.go
@@ -1,62 +1,0 @@
-package manager
-
-import (
-	"context"
-	"errors"
-	"log"
-	"regexp"
-	"time"
-
-	"cloud.google.com/go/storage"
-	"google.golang.org/api/iterator"
-
-	"github.com/m-lab/annotation-service/api"
-)
-
-// Init extracts the filenames from downloader bucket
-// It also searches the latest Geolite2 files available in GCS.
-// It will also set LatestDatasetDate as the date of lastest dataset.
-// If it encounters an error, it will halt the program.
-func Init() error {
-	ctx := context.Background()
-	client, err := storage.NewClient(ctx)
-	if err != nil {
-		return err
-	}
-	prospectiveFiles := client.Bucket(api.MaxmindBucketName).Objects(ctx, &storage.Query{Prefix: api.MaxmindPrefix})
-	filename := ""
-	for file, err := prospectiveFiles.Next(); err != iterator.Done; file, err = prospectiveFiles.Next() {
-		if err != nil {
-			return err
-		}
-		DatasetNames = append(DatasetNames, file.Name)
-		if file.Name > filename && GeoLite2Regex.MatchString(file.Name) {
-			filename = file.Name
-		}
-	}
-	if err != nil {
-		log.Println(err)
-	}
-
-	// Now set the lastest dataset
-	date, err := ExtractDateFromFilename(filename)
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-	LatestDatasetDate = date
-	return nil
-}
-
-// ExtractDateFromFilename return the date for a filename like
-// gs://downloader-mlab-oti/Maxmind/2017/05/08/20170508T080000Z-GeoLiteCity.dat.gz
-// TODO move this to maxmind package
-// TODO - actually, this now seems to be dead code.  But probably needed again soon, so leaving it here.
-func ExtractDateFromFilename(filename string) (time.Time, error) {
-	re := regexp.MustCompile(`[0-9]{8}T`)
-	filedate := re.FindAllString(filename, -1)
-	if len(filedate) != 1 {
-		return time.Time{}, errors.New("cannot extract date from input filename")
-	}
-	return time.Parse(time.RFC3339, filedate[0][0:4]+"-"+filedate[0][4:6]+"-"+filedate[0][6:8]+"T00:00:00Z")
-}
